Logout Controller Delete - Add token authentication in Swagger - Separation of user - auth logic
