--- src/controllers/logoutController.ts
@@ -1,5 +1,0 @@
-import { Request, Response } from 'express';
-
-export const logoutUser = async (_req: Request, res: Response) => {
-    res.status(200).json({ message: 'Logout exitoso' });
-};
