+++ src/routes/userRoutes.ts
@@ -1,12 +1,8 @@
 import express from 'express';
-import { loginUser, registerUser, getUsers, getUserById } from '../controllers/userController';
-import { logoutUser } from '../controllers/logoutController';
+import { getUsers, getUserById } from '../controllers/userController';
 
 const userRoutes = express.Router();
 
-userRoutes.post('/login', loginUser);
-userRoutes.post('/register', registerUser);
-userRoutes.post('/logout', logoutUser)
 userRoutes.get('/', getUsers);
 userRoutes.get('/:id', getUserById);
