+++ src/swagger/user/routes.ts
@@ -6,152 +6,7 @@
 
 /**
  * @swagger
- * /auth/login:
- *   post:
- *     summary: Inicia sesión de un usuario
- *     description: Autentica a un usuario en el sistema con el correo electrónico y la contraseña proporcionados.
- *     tags:
- *       - user
- *     requestBody:
- *       required: true
- *       content:
- *         application/json:
- *           schema:
- *             $ref: '#/components/schemas/UserLogin'
- *     responses:
- *       200:
- *         description: Inicio de sesión exitoso
- *         content:
- *           application/json:
- *             schema:
- *               type: object
- *               properties:
- *                 message:
- *                   type: string
- *                   example: "Inicio de sesión exitoso"
- *                 user:
- *                   $ref: '#/components/schemas/User'
- *       400:
- *         description: Credenciales incorrectas
- *         content:
- *           application/json:
- *             schema:
- *               type: object
- *               properties:
- *                 message:
- *                   type: string
- *                   example: "Contraseña incorrecta"
- *       500:
- *         description: Error interno del servidor
- *         content:
- *           application/json:
- *             schema:
- *               type: object
- *               properties:
- *                 message:
- *                   type: string
- *                   example: "Error al iniciar sesión"
- *                 error:
- *                   type: string
- *                   example: "Error específico del servidor"
- */
-
-/**
- * @swagger
- * /auth/register:
- *   post:
- *     summary: Registra un nuevo usuario
- *     description: Registra un nuevo usuario en el sistema con los datos proporcionados.
- *     tags:
- *       - user
- *     requestBody:
- *       required: true
- *       content:
- *         application/json:
- *           schema:
- *             $ref: '#/components/schemas/UserDto'
- *     responses:
- *       201:
- *         description: Usuario registrado exitosamente
- *         content:
- *           application/json:
- *             schema:
- *               type: object
- *               properties:
- *                 message:
- *                   type: string
- *                   example: "Usuario registrado"
- *                 user:
- *                   $ref: '#/components/schemas/User'
- *       400:
- *         description: Solicitud mal formada o con datos faltantes
- *         content:
- *           application/json:
- *             schema:
- *               type: object
- *               properties:
- *                 message:
- *                   type: string
- *                   example: "Faltan campos obligatorios"
- *       500:
- *         description: Error interno del servidor
- *         content:
- *           application/json:
- *             schema:
- *               type: object
- *               properties:
- *                 message:
- *                   type: string
- *                   example: "Error al registrar el usuario"
- *                 error:
- *                   type: string
- *                   example: "Error específico del servidor"
- */
-
-/**
- * @swagger
- * /auth/logout:
- *   post:
- *     summary: Cierra la sesión de un usuario
- *     description: Finaliza la sesión del usuario autenticado.
- *     tags:
- *       - user
- *     responses:
- *       200:
- *         description: Cierre de sesión exitoso
- *         content:
- *           application/json:
- *             schema:
- *               type: object
- *               properties:
- *                 message:
- *                   type: string
- *                   example: "Sesión cerrada exitosamente"
- *       403:
- *         description: Token no proporcionado o inválido
- *         content:
- *           application/json:
- *             schema:
- *               type: object
- *               properties:
- *                 message:
- *                   type: string
- *                   example: "Acceso no autorizado"
- *       500:
- *         description: Error interno del servidor
- *         content:
- *           application/json:
- *             schema:
- *               type: object
- *               properties:
- *                 message:
- *                   type: string
- *                   example: "Error al cerrar sesión"
- */
-
-/**
- * @swagger
- * /auth:
+ * /user:
  *   get:
  *     summary: Obtiene la lista de todos los usuarios
  *     description: Recupera todos los usuarios registrados en el sistema.
@@ -180,7 +35,7 @@
 
 /**
  * @swagger
- * /auth/{id}:
+ * /user/{id}:
  *   get:
  *     summary: Obtiene un usuario por ID
  *     description: Recupera los detalles de un usuario específico utilizando su ID.
